--- PillTip/BE/test/RAG/data/expression_mappings.json
@@ -1,55 +1,0 @@
-{
-  "속이 더부룩": "소화불량 ",
-  "배가 더부룩": "소화불량 ",
-  "소화가 안 돼": "소화불량 ",
-  "체했": "소화불량 ",
-  "속이 울렁": "메스꺼움 ",
-  "토할 것": "구토 ",
-  "머리가 띵": "두통 ",
-  "어지러": "현기증 ",
-  "몸이 쑤셔": "근육통 ",
-  "가슴이 답답": "호흡곤란 ",
-  "숨쉬기 힘들": "호흡곤란 ",
-  "입맛이 없": "식욕부진 ",
-  "기운이 없": "피로 ",
-  "맥이 없": "무기력 ",
-  "가슴이 철렁": "심계항진 ",
-  "심장이 터질 것": "심계항진 ",
-  "머리가 깨질 것": "두통 ",
-  "하늘이 노래": "현기증 ",
-  "하늘이 노랗게": "현기증 ",
-  "헉헉거려": "호흡곤란 ",
-  "몸이 으슬으슬": "오한 ",
-  "몸이 으슬": "오한 ",
-  "허리가 끊어질 것": "요통 ",
-  "토 나와": "구토 ",
-  "트림": "소화불량 ",
-  "배가 우리": "복통 ",
-  "땀이 줄줄": "땀샘질환 ",
-  "식은땀": "땀샘질환 ",
-  "속이 쓰려": "속쓰림 ",
-  "목이 따끔": "인후통 ",
-  "목이 타들": "갈증 ",
-  "쿨럭쿨럭": "기침 ",
-  "콜록콜록": "기침 ",
-  "뒷골": "후두통 ",
-  "신물": "위산 역류 ",
-  "손이 시": "수족냉증 ",
-  "발이 시": "수족냉증 ",
-  "손발이 시": "수족냉증 ",
-  "손발이 저": "손발 저림 ",
-  "손이 저": "손발 저림 ",
-  "발이 저": "손발 저림 ",
-  "손끝이 저": "손발 저림 ",
-  "눈이 흐릿": "시야 흐림 ",
-  "초점이 안 맞": "시야 흐림 ",
-  "초점이 안맞": "시야 흐림 ",
-  "초점이 잘 안맞": "시야 흐림 ",
-  "초점이 잘 안 맞": "시야 흐림 ",
-  "눈이 잘 안 보": "시야 흐림 ",
-  "눈이 안 보": "시야 흐림 ",
-  "눈이 안보": "시야 흐림 ",
-  "땀이 많이": "땀샘질환 ",
-  "몸이 뜨거": "발열 ",
-  "몸이 뜨겁": "발열 "
-}
