--- PillTip/BE/test/RAG/data/requirements.txt
@@ -1,6 +1,0 @@
-pymilvus>=2.4.0
-scikit-learn
-openai
-numpy
-tqdm
-
