+++ PillTip/BE/test/RAG/milvus/requirements.txt
@@ -1,0 +1,6 @@
+pymilvus>=2.4.0
+scikit-learn
+openai
+numpy
+tqdm
+
